+++ src/lib/Prefs/PrefsKeys.js
@@ -593,6 +593,18 @@ var PrefsKeys = class
 
         this.setKey(
             'behavior',
+            'overlay-key',
+            'GtkSwitch',
+            this._shellVersion >= 40,
+            {
+                default: true,
+                minimal: true,
+                superminimal: true,
+            }
+        );
+
+        this.setKey(
+            'behavior',
             'double-super-to-appgrid',
             'GtkSwitch',
             this._shellVersion >= 40,